feat: 파티-파티멤버-룸 엔티티 및 비즈니스 로직 작성 #BACK-124 Related to: #124, #126, #128 # 164

+++ backend/src/main/java/signiel/heartsigniel/model/user/User.java
@@ -30,7 +30,7 @@ public class User{
     private Date birth;
     private String phoneNumber;
     private boolean isBlack;
-    private int rank;
+    private Long rank;
     private int meetingCount;
     private String profileImageSrc;
     @NotBlank(message = "직업을 입력해주세요.")
